Fix struct type/layout regressions from renames
Restore original codegen for functions broken by type changes made
alongside the symbol/field renames:

- FighterData/NameTagData sd_count: revert u32 back to u16 + padding so
  loads/stores and fn_80161C90's u16* arg match again.
- gmm_x1868 event_match_records: size is NUM_EVENT_MATCHES (51), not
  NUM_EVENT_MATCHES - 1; the short array shifted following fields.
- gm_801BEBF8: undo an extra pointer indirection; read the byte directly
  from selectedEventMatch->x14.
- gm_801BAB40: restore the int-arg signature and raw bit assignments
  (the renamed PLAYER_FLAGS_* masks were shifted by the wrong amounts).

+++ src/melee/gm/gm_1601.c
@@ -1124,7 +1124,7 @@ struct gm_stats {
     /* 0x40 */ u32 unk40;
 };
 
-void fn_80161C90(MatchEnd* arg0, int slot, u32* arg2)
+void fn_80161C90(MatchEnd* arg0, int slot, u16* arg2)
 {
     MatchPlayerData* p = &arg0->player_standings[slot];
     struct gm_stats* s = (struct gm_stats*) arg2;
